refactor: PPS-41 extracted MockServer for tests inside a trait useful for Cucumber test definitions
